requirements.txt
fastapi
uvicorn
grpcio
yandexcloud
python-dotenv